Remove inaccurate documentation
This is completely untrue. The reason that the tests were failing was
because the path to one of the setup scripts needs to be modified to
reference the hoisted dependency. Ideally it would just reference the
file using npm's node_modules upward cascading traversal so that it's
resilient to being run in multiple contexts.